requirements.txt
click>=8.0.0
boto3>=1.26.0
psutil>=5.9.0
redis>=4.5.0
pyyaml>=6.0
rich>=13.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
requests>=2.28.0
colorama>=0.4.6
